feat(settings): move transcript normalization into AI & Processing
Normalization corrects Whisper output for words, phrases, numbers, and
sponsor names alike, so it now lives in settings as its own collapsible
section next to Transcription, with help text saying so. The Sponsors
page drops its tab bar and keeps only the sponsor list.
